fixed circular deps with asset-swapper market operations stuff
