+++ src/asset-swapper/utils/market_operation_utils/identity_fill_adjustor.ts
@@ -1,6 +1,4 @@
-import { MarketOperation } from '../../types';
-
-import { Fill, FillAdjustor } from './types';
+import { MarketOperation, Fill, FillAdjustor } from '../../types';
 
 export class IdentityFillAdjustor implements FillAdjustor {
     public adjustFills(_side: MarketOperation, fills: Fill[]): Fill[] {
