+++ src/asset-swapper/utils/market_operation_utils/sampler_contract_operation.ts
@@ -3,7 +3,8 @@ import { BigNumber, decodeBytesAsRevertError, logUtils } from '@0x/utils';
 
 import { ERC20BridgeSamplerContract } from '../../../wrappers';
 
-import { ERC20BridgeSource, FillData, SourceQuoteOperation } from './types';
+import { ERC20BridgeSource, FillData } from '../../types';
+import { SourceQuoteOperation } from './types';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any -- TODO: fix me!
 export type Parameters<T> = T extends (...args: infer TArgs) => any ? TArgs : never;
