+++ test/asset-swapper/comparison_price_test.ts
@@ -4,14 +4,10 @@ import * as _ from 'lodash';
 import 'mocha';
 
 import { SOURCE_FLAGS } from '../../src/asset-swapper';
-import { MarketOperation } from '../../src/asset-swapper/types';
+import { MarketOperation, ERC20BridgeSource } from '../../src/asset-swapper/types';
 import { getComparisonPrices } from '../../src/asset-swapper/utils/market_operation_utils/comparison_price';
 import { SourceFilters } from '../../src/asset-swapper/utils/market_operation_utils/source_filters';
-import {
-    DexSample,
-    ERC20BridgeSource,
-    MarketSideLiquidity,
-} from '../../src/asset-swapper/utils/market_operation_utils/types';
+import { DexSample, MarketSideLiquidity } from '../../src/asset-swapper/utils/market_operation_utils/types';
 
 import { chaiSetup } from './utils/chai_setup';
