+++ test/asset-swapper/exchange_proxy_swap_quote_consumer_test.ts
@@ -26,14 +26,12 @@ import {
     MarketBuySwapQuote,
     MarketOperation,
     MarketSellSwapQuote,
-} from '../../src/asset-swapper/types';
-import {
     ERC20BridgeSource,
     Fill,
     NativeFillData,
     OptimizedLimitOrder,
     OptimizedMarketOrder,
-} from '../../src/asset-swapper/utils/market_operation_utils/types';
+} from '../../src/asset-swapper/types';
 
 import { chaiSetup } from './utils/chai_setup';
 import { getRandomAmount, getRandomSignature } from './utils/utils';
