+++ test/asset-swapper/utils/market_operations_utils/path_optimizer_test.ts
@@ -9,13 +9,8 @@ import {
     SignatureType,
 } from '../../../../src/asset-swapper';
 import { PathOptimizer } from '../../../../src/asset-swapper/utils/market_operation_utils/path_optimizer';
-import {
-    DexSample,
-    ERC20BridgeSource,
-    FeeSchedule,
-    FillData,
-    MultiHopFillData,
-} from '../../../../src/asset-swapper/utils/market_operation_utils/types';
+import { ERC20BridgeSource, FeeSchedule, FillData } from '../../../../src/asset-swapper/types';
+import { DexSample, MultiHopFillData } from '../../../../src/asset-swapper/utils/market_operation_utils/types';
 import { BigNumber } from '@0x/utils';
 import { chaiSetup } from '../chai_setup';
 import { PathPenaltyOpts } from '../../../../src/asset-swapper/utils/market_operation_utils/path';
